Backend ver 0.3.1
- Update Docker Compose (add build and prod variants)
- Added bash and ps scripts to run app
- Updates and optimization in routers
- Update user model

+++ backend/app/database/init_db.py
@@ -1,13 +1,17 @@
 from app.database.base import Base
 from app.database.engine import engine
+from app.config import DATABASE_DROP_ALL
 
 
 def init_db():
     try:
-        Base.metadata.drop_all(bind=engine)
+        if DATABASE_DROP_ALL:
+            print("[INFO] Dropping all tables...")
+            Base.metadata.drop_all(bind=engine)
+
         Base.metadata.create_all(bind=engine)
         print("[SUCCESS] Tables created!")
-        print(f"Tables list: {list(Base.metadata.tables.keys())}")
+        print(f"[INFO] Tables list: {list(Base.metadata.tables.keys())}")
     except Exception as e:
         print(f"[ERROR] Error while creating tables {str(e)}")
     finally:
